share remaining validation with pass throughs for additional platform/profile specific validation
